경고 수정

+++ Common/DefaultPawn.cpp
@@ -89,7 +89,7 @@ void DefaultPawn::OnInputProcess(const Keyboard::State& KeyState, const Keyboard
 	else if (MouseTracker.leftButton == Mouse::ButtonStateTracker::PRESSED)
 	{		
 		Math::Vector3 org,dir;
-		D3DRenderManager::Instance->CreateMousePickingRay(MouseState.x, MouseState.y, org, dir);			
+		D3DRenderManager::Instance->CreateMousePickingRay((float)MouseState.x, (float)MouseState.y, org, dir);
 		D3DRenderManager::Instance->AddDebugDrawLine(org, dir * 100000000,false,Math::Vector3(1.0f,1.0f,0.0f),10.0f);
 		
 		std::list<RayResult> result;
